requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.7
django-phonenumber-field==7.0.2
kombu==5.2.4
phonenumbers==8.13.7
prompt-toolkit==3.0.38
python-telegram==0.18.0
pytz==2023.2
redis==4.5.3
six==1.16.0
sqlparse==0.4.3
telegram-text==0.1.1
vine==5.0.0
wcwidth==0.2.6
